Move CSV query builders to csv_builder.py
